create new food, image upload

+++ client/pages/api/products/[id].js
@@ -4,7 +4,8 @@ import { ServiceError } from '../../../lib/errors';
 
 
 export default async function handler(req, res) {
-    const { method, query: { id } } = req;
+    const { method, query: { id }, cookies } = req;
+    const token = cookies.token;
 
     await dbConnect();
     if (method === 'GET') {
@@ -16,6 +17,9 @@ export default async function handler(req, res) {
         }
     }
     if (method === 'POST') {
+        if (!token || token !== process.env.TOKEN) {
+            return res.status(401).json('You are unauthorized to perform this action')
+        }
         try {
             const product = await Product.create(req.body)
             res.status(201).json(product)
@@ -25,6 +29,9 @@ export default async function handler(req, res) {
     }
 
     if (method === 'DELETE') {
+        if (!token || token !== process.env.TOKEN) {
+            return res.status(401).json('You are unauthorized to perform this action')
+        }
         try {
             await Product.findByIdAndDelete(id);
             res.status(200).json({message: 'successfully deleted.'})
